Read the ASD files and return as SummarizedExperiment object.

Read ASD Spectral Files and Create a SummarizedExperiment Object

Arguments

path

A character string specifying the path to a directory containing the ".asd" files or a single ".asd" file.

Value

A `SummarizedExperiment` object containing:

assaysA list with a single assay matrix named counts, where rows represent wavelengths and columns represent different samples.

colDataA `DataFrame` with sample metadata, including the sample names.

rowDataA `DataFrame` with wavelength information.

Details

This function will read a single or multiple ASD files and extracts the spectral data and associated metadata, combines them into a "SummarizedExperiment" object. This object contains the spectral data as an assay matrix and metadata about the samples and wavelengths.

This function uses 2 functions such as "read.asd" and "extract.metadata" from the package "FieldSpectra" to read the data and metadata from the given files. This function will then uses the data from the files and make a SummarizedExperiment object out of it which can be used for further data analysis.

Author

Methun George, Dr.Steffen Neumann

Examples

# path <- "path/to/your/asd/files"
# se <- read_asd_summar(path)
# se
# colnames(se)
# rownames(se)