Read the SIG files and return as SummarizedExperiment object.

Read .sig Files and Create SummarizedExperiment Object

Arguments

path

A character vector of file paths or a directory path. If a directory path is provided, all .sig files within the directory are read. If a file path is provided, only that file is read.

Value

A SummarizedExperiment object where:

assays

A matrix of reflectance data, with wavelengths as row names and sample names as column names.

colData

A DataFrame with sample names.

rowData

A DataFrame with wavelengths.

Details

This function reads .sig files from a specified directory or a file path, extracts wavelength and reflectance data, and combines them into a SummarizedExperiment object. Each .sig file is expected to contain reflectance data and a sample name in its content.

Each .sig file is read line by line. The function expects each file to contain:

A line starting with "name=" indicating the sample name.

A section starting with "data=" followed by tabular data where the first column is wavelengths and the fourth column is reflectance.

The combined data from all files is merged by wavelengths. Missing data is handled by filling with NA.

Author

Methun George, Dr.Steffen Neumann

Examples

# Example usage:
path <- "/path/to/sig/files"
se <- read_summarizedexperiment_asd(path)
#> Error in if (length(path) == 1 && file.info(path)$isdir) {    asd_files <- list.files(path = path, pattern = "\\.asd$",         full.names = TRUE)} else {    asd_files <- path}: missing value where TRUE/FALSE needed
print(se)
#> Error in eval(expr, envir, enclos): object 'se' not found